introduce ezLib with funtions to import userModules
Previously adding a homeConfiguration to a host was only possible
via setting `ezConfigs.darwin.hosts.<name>.userHomeModules` or
`ezConfigs.nixos.hosts.<name>.userHomeModules` at the flake level.
Importing a userModule inside a `nixosModule` or `darwinModule` was
not possible.

This commit introduces
  - `ezLib.nixosImportUserModule`
  - `ezLib.darwinImportUserModule`
functions to allow this usage. `ezLib` is part of `specialArgs` and
`extraSpecialArgs` and thus available in modules by default.

Usage example for a `nixosModules/alice.nix` module:
```nix
{ezLib, ...}: {
  imports = [
    (ezLib.nixosImportUserModule "alice")
  ];

  # User configuration, as given in nixos option example for users.users
  users.users.alice = {
    createHome = true;
    description = "Alice Q. User";
    extraGroups = [
      "wheel"
    ];
    group = "users";
    home = "/home/alice";
    shell = "/bin/sh";
    uid = 1234;
  };
}
```

Usage of `ezLib.darwinImportUserModule` is analogous.